+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # row to object
 ## revisions
 
+@@2020-06-022
+- added the $$IF statement with (see object-to-object.spec for examples)
+
 ## 2020-05-31
 - better message when 'TypeError'
 - add padStart